@@ -4,7 +4,7 @@ filesystem (e.g.: local, S3, GCS). It uses native json to handle the JSON file.
 import json
 from copy import deepcopy
 from pathlib import PurePosixPath
-from typing import Any, Dict
+from typing import Any
 
 import fsspec
 from kedro.io.core import Version, get_filepath_str, get_protocol_and_path
@@ -46,16 +46,17 @@ class JSONDataset(AbstractVersionedDataset[Any, Any]):
 
     """
 
-    DEFAULT_SAVE_ARGS: Dict[str, Any] = {"indent": 2}
+    DEFAULT_SAVE_ARGS: dict[str, Any] = {"indent": 2}
 
     def __init__(  # noqa: PLR0913
         self,
+        *,
         filepath: str,
-        save_args: Dict[str, Any] = None,
+        save_args: dict[str, Any] = None,
         version: Version = None,
-        credentials: Dict[str, Any] = None,
-        fs_args: Dict[str, Any] = None,
-        metadata: Dict[str, Any] = None,
+        credentials: dict[str, Any] = None,
+        fs_args: dict[str, Any] = None,
+        metadata: dict[str, Any] = None,
     ) -> None:
         """Creates a new instance of ``JSONDataset`` pointing to a concrete JSON file
         on a specific filesystem.
@@ -116,7 +117,7 @@ class JSONDataset(AbstractVersionedDataset[Any, Any]):
         self._fs_open_args_load = _fs_open_args_load
         self._fs_open_args_save = _fs_open_args_save
 
-    def _describe(self) -> Dict[str, Any]:
+    def _describe(self) -> dict[str, Any]:
         return {
             "filepath": self._filepath,
             "protocol": self._protocol,
