@@ -5,7 +5,7 @@ import logging
 from copy import deepcopy
 from io import BytesIO
 from pathlib import PurePosixPath
-from typing import Any, Dict, Union
+from typing import Any, Union
 
 import fsspec
 import pandas as pd
@@ -23,8 +23,8 @@ logger = logging.getLogger(__name__)
 
 class ExcelDataset(
     AbstractVersionedDataset[
-        Union[pd.DataFrame, Dict[str, pd.DataFrame]],
-        Union[pd.DataFrame, Dict[str, pd.DataFrame]],
+        Union[pd.DataFrame, dict[str, pd.DataFrame]],
+        Union[pd.DataFrame, dict[str, pd.DataFrame]],
     ]
 ):
     """``ExcelDataset`` loads/saves data from/to a Excel file using an underlying
@@ -110,14 +110,15 @@ class ExcelDataset(
 
     def __init__(  # noqa: PLR0913
         self,
+        *,
         filepath: str,
         engine: str = "openpyxl",
-        load_args: Dict[str, Any] = None,
-        save_args: Dict[str, Any] = None,
+        load_args: dict[str, Any] = None,
+        save_args: dict[str, Any] = None,
         version: Version = None,
-        credentials: Dict[str, Any] = None,
-        fs_args: Dict[str, Any] = None,
-        metadata: Dict[str, Any] = None,
+        credentials: dict[str, Any] = None,
+        fs_args: dict[str, Any] = None,
+        metadata: dict[str, Any] = None,
     ) -> None:
         """Creates a new instance of ``ExcelDataset`` pointing to a concrete Excel file
         on a specific filesystem.
@@ -202,7 +203,7 @@ class ExcelDataset(
             self._save_args.pop("storage_options", None)
             self._load_args.pop("storage_options", None)
 
-    def _describe(self) -> Dict[str, Any]:
+    def _describe(self) -> dict[str, Any]:
         return {
             "filepath": self._filepath,
             "protocol": self._protocol,
@@ -212,7 +213,7 @@ class ExcelDataset(
             "version": self._version,
         }
 
-    def _load(self) -> Union[pd.DataFrame, Dict[str, pd.DataFrame]]:
+    def _load(self) -> Union[pd.DataFrame, dict[str, pd.DataFrame]]:
         load_path = str(self._get_load_path())
         if self._protocol == "file":
             # file:// protocol seems to misbehave on Windows
@@ -226,7 +227,7 @@ class ExcelDataset(
             load_path, storage_options=self._storage_options, **self._load_args
         )
 
-    def _save(self, data: Union[pd.DataFrame, Dict[str, pd.DataFrame]]) -> None:
+    def _save(self, data: Union[pd.DataFrame, dict[str, pd.DataFrame]]) -> None:
         output = BytesIO()
         save_path = get_filepath_str(self._get_save_path(), self._protocol)
 
@@ -261,7 +262,7 @@ class ExcelDataset(
         filepath = get_filepath_str(self._filepath, self._protocol)
         self._fs.invalidate_cache(filepath)
 
-    def _preview(self, nrows: int = 40) -> Dict:
+    def _preview(self, nrows: int = 40) -> dict:
         # Create a copy so it doesn't contaminate the original dataset
         dataset_copy = self._copy()
         dataset_copy._load_args["nrows"] = nrows
