@@ -3,7 +3,7 @@ file using an underlying filesystem (e.g.: local, S3, GCS). It loads the JSON in
 plotly figure.
 """
 from copy import deepcopy
-from typing import Any, Dict
+from typing import Any
 
 import pandas as pd
 import plotly.express as px
@@ -68,14 +68,15 @@ class PlotlyDataset(JSONDataset):
 
     def __init__(  # noqa: PLR0913
         self,
+        *,
         filepath: str,
-        plotly_args: Dict[str, Any],
-        load_args: Dict[str, Any] = None,
-        save_args: Dict[str, Any] = None,
+        plotly_args: dict[str, Any],
+        load_args: dict[str, Any] = None,
+        save_args: dict[str, Any] = None,
         version: Version = None,
-        credentials: Dict[str, Any] = None,
-        fs_args: Dict[str, Any] = None,
-        metadata: Dict[str, Any] = None,
+        credentials: dict[str, Any] = None,
+        fs_args: dict[str, Any] = None,
+        metadata: dict[str, Any] = None,
     ) -> None:
         """Creates a new instance of ``PlotlyDataset`` pointing to a concrete JSON file
         on a specific filesystem.
@@ -113,7 +114,14 @@ class PlotlyDataset(JSONDataset):
             metadata: Any arbitrary metadata.
                 This is ignored by Kedro, but may be consumed by users or external plugins.
         """
-        super().__init__(filepath, load_args, save_args, version, credentials, fs_args)
+        super().__init__(
+            filepath=filepath,
+            load_args=load_args,
+            save_args=save_args,
+            version=version,
+            credentials=credentials,
+            fs_args=fs_args,
+        )
         self._plotly_args = plotly_args
 
         _fs_args = deepcopy(fs_args) or {}
@@ -126,7 +134,7 @@ class PlotlyDataset(JSONDataset):
 
         self.metadata = metadata
 
-    def _describe(self) -> Dict[str, Any]:
+    def _describe(self) -> dict[str, Any]:
         return {**super()._describe(), "plotly_args": self._plotly_args}
 
     def _save(self, data: pd.DataFrame) -> None:
