@@ -1,7 +1,7 @@
 """SparkStreamingDataset to load and save a PySpark Streaming DataFrame."""
 from copy import deepcopy
 from pathlib import PurePosixPath
-from typing import Any, Dict
+from typing import Any
 
 from pyspark.sql import DataFrame
 from pyspark.sql.utils import AnalysisException
@@ -42,10 +42,11 @@ class SparkStreamingDataset(AbstractDataset):
 
     def __init__(
         self,
+        *,
         filepath: str = "",
         file_format: str = "",
-        save_args: Dict[str, Any] = None,
-        load_args: Dict[str, Any] = None,
+        save_args: dict[str, Any] = None,
+        load_args: dict[str, Any] = None,
     ) -> None:
         """Creates a new instance of SparkStreamingDataset.
 
@@ -94,7 +95,7 @@ class SparkStreamingDataset(AbstractDataset):
             if isinstance(self._schema, dict):
                 self._schema = SparkDataset._load_schema_from_file(self._schema)
 
-    def _describe(self) -> Dict[str, Any]:
+    def _describe(self) -> dict[str, Any]:
         """Returns a dict that describes attributes of the dataset."""
         return {
             "filepath": self._fs_prefix + str(self._filepath),
