@@ -5,9 +5,10 @@ and decode videos and OpenCV VideoWriter to encode and write video.
 import itertools
 import tempfile
 from collections import abc
+from collections.abc import Generator, Sequence
 from copy import deepcopy
 from pathlib import Path, PurePosixPath
-from typing import Any, Dict, Generator, Optional, Sequence, Tuple, Union
+from typing import Any, Optional, Union
 
 import cv2
 import fsspec
@@ -54,7 +55,7 @@ class AbstractVideo(abc.Sequence):
         raise NotImplementedError()
 
     @property
-    def size(self) -> Tuple[int, int]:
+    def size(self) -> tuple[int, int]:
         """Get the resolution of the video"""
         raise NotImplementedError()
 
@@ -84,7 +85,7 @@ class FileVideo(AbstractVideo):
         return self._cap.get(cv2.CAP_PROP_FPS)
 
     @property
-    def size(self) -> Tuple[int, int]:
+    def size(self) -> tuple[int, int]:
         width = int(self._cap.get(cv2.CAP_PROP_FRAME_WIDTH))
         height = int(self._cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
         return width, height
@@ -148,7 +149,7 @@ class SequenceVideo(AbstractVideo):
         return self._fps
 
     @property
-    def size(self) -> Tuple[int, int]:
+    def size(self) -> tuple[int, int]:
         return self._size
 
     def __getitem__(self, index: Union[int, slice]):
@@ -183,7 +184,7 @@ class GeneratorVideo(AbstractVideo):
         return self._fps
 
     @property
-    def size(self) -> Tuple[int, int]:
+    def size(self) -> tuple[int, int]:
         return self._size
 
     def __getitem__(self, index: Union[int, slice]):
@@ -224,9 +225,11 @@ class VideoDataset(AbstractDataset[AbstractVideo, AbstractVideo]):
         >>> from kedro_datasets.video import VideoDataset
         >>> import numpy as np
         >>>
-        >>> video = VideoDataset(filepath="/video/file/path.mp4").load()
+        >>> video = VideoDataset(
+        ...     filepath="https://storage.googleapis.com/gtv-videos-bucket/sample/ForBiggerBlazes.mp4"
+        ... ).load()
         >>> frame = video[0]
-        >>> np.sum(np.asarray(frame))
+        >>> assert isinstance(np.sum(np.asarray(frame)), np.uint64)
 
 
     Example creating a video from numpy frames using Python API:
@@ -243,7 +246,7 @@ class VideoDataset(AbstractDataset[AbstractVideo, AbstractVideo]):
         ...     imgs.append(Image.fromarray(frame))
         ...     frame -= 1
         ...
-        >>> video = VideoDataset("my_video.mp4")
+        >>> video = VideoDataset(filepath="my_video.mp4")
         >>> video.save(SequenceVideo(imgs, fps=25))
 
 
@@ -261,18 +264,19 @@ class VideoDataset(AbstractDataset[AbstractVideo, AbstractVideo]):
         ...         yield Image.fromarray(frame)
         ...         frame -= 1
         ...
-        >>> video = VideoDataset("my_video.mp4")
+        >>> video = VideoDataset(filepath="my_video.mp4")
         >>> video.save(GeneratorVideo(gen(), fps=25, length=None))
 
     """
 
     def __init__(  # noqa: PLR0913
         self,
+        *,
         filepath: str,
         fourcc: Optional[str] = "mp4v",
-        credentials: Dict[str, Any] = None,
-        fs_args: Dict[str, Any] = None,
-        metadata: Dict[str, Any] = None,
+        credentials: dict[str, Any] = None,
+        fs_args: dict[str, Any] = None,
+        metadata: dict[str, Any] = None,
     ) -> None:
         """Creates a new instance of VideoDataset to load / save video data for given filepath.
 
@@ -360,7 +364,7 @@ class VideoDataset(AbstractDataset[AbstractVideo, AbstractVideo]):
         finally:
             writer.release()
 
-    def _describe(self) -> Dict[str, Any]:
+    def _describe(self) -> dict[str, Any]:
         return {"filepath": self._filepath, "protocol": self._protocol}
 
     def _exists(self) -> bool:
