@@ -73,7 +73,7 @@ class TestDeltaTableDataset:
 
     def test_versioning(self, filepath, dummy_df):
         """Test loading different versions."""
-        deltatable_dataset_from_path = DeltaTableDataset(filepath)
+        deltatable_dataset_from_path = DeltaTableDataset(filepath=filepath)
         deltatable_dataset_from_path.save(dummy_df)
         assert deltatable_dataset_from_path.get_loaded_version() == 0
         new_df = pd.DataFrame({"col1": [0, 0], "col2": [1, 1], "col3": [2, 2]})
@@ -81,14 +81,14 @@ class TestDeltaTableDataset:
         assert deltatable_dataset_from_path.get_loaded_version() == 1
 
         deltatable_dataset_from_path0 = DeltaTableDataset(
-            filepath, load_args={"version": 0}
+            filepath=filepath, load_args={"version": 0}
         )
         version_0 = deltatable_dataset_from_path0.load()
         assert deltatable_dataset_from_path0.get_loaded_version() == 0
         assert_frame_equal(dummy_df, version_0)
 
         deltatable_dataset_from_path1 = DeltaTableDataset(
-            filepath, load_args={"version": 1}
+            filepath=filepath, load_args={"version": 1}
         )
         version_1 = deltatable_dataset_from_path1.load()
         assert deltatable_dataset_from_path1.get_loaded_version() == 1
@@ -108,7 +108,7 @@ class TestDeltaTableDataset:
 
     def test_describe(self, filepath):
         """Test the describe method."""
-        deltatable_dataset_from_path = DeltaTableDataset(filepath)
+        deltatable_dataset_from_path = DeltaTableDataset(filepath=filepath)
         desc = deltatable_dataset_from_path._describe()
         assert desc["filepath"] == filepath
         assert desc["version"] is None
@@ -152,7 +152,7 @@ class TestDeltaTableDataset:
         """Test write mode not supported."""
         pattern = "Write mode unsupported is not supported"
         with pytest.raises(DatasetError, match=pattern):
-            DeltaTableDataset(filepath, save_args={"mode": "unsupported"})
+            DeltaTableDataset(filepath=filepath, save_args={"mode": "unsupported"})
 
     def test_metadata(self, deltatable_dataset_from_path, dummy_df):
         """Test metadata property exists and return a metadata object."""
